pyproject.toml
[build-system]
requires = [
    "setuptools>=69.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "ansible-directory-helper"
version = "0.1.0"
authors = [
    {name="Brian St. Hilaire", email="brian.st-hilaire@sanctus-tech.com"}
]
description = "A small helper library to manage local directories for Ansible"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml",
    "ansible-core",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Framework :: Ansible",
    "Programming Language :: Python :: 3",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
]

[project.urls]
Homepage = "https://github.com/saint-hilaire/ansible-directory-helper"
Issues = "https://github.com/saint-hilaire/ansible-directory-helper/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
